refactor: MyBook でまとめた
