+++ core/data/src/main/java/app/kaito_dogi/mybrary/core/data/dto/MyBookDto.kt
@@ -4,9 +4,9 @@ data class MyBookDto(
   val myBookId: String,
   val title: String,
   val imageUrl: String,
-  val isbn: String,
-  val publisher: String,
   val authors: String,
+  val publisher: String,
+  val isbn: String,
   val genre: Int,
   val isPinned: Boolean,
   val isFavorite: Boolean,
